+++ src/main/java/com/example/whatseoul/controller/mypage/MyPageController.java
@@ -1,7 +1,7 @@
 package com.example.whatseoul.controller.mypage;
 
 import com.example.whatseoul.entity.User;
-import com.example.whatseoul.respository.user.UserRepository;
+import com.example.whatseoul.repository.user.UserRepository;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
