+++ src/main/java/com/example/whatseoul/repository/TestRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.example.whatseoul.respository;
+package com.example.whatseoul.repository;
 
 
 import org.springframework.stereotype.Repository;
