+++ src/main/java/com/example/whatseoul/repository/cityData/PopulationForecastRepository.java
@@ -1,11 +1,9 @@
-package com.example.whatseoul.respository.cityData;
+package com.example.whatseoul.repository.cityData;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
-import com.example.whatseoul.entity.Area;
 import com.example.whatseoul.entity.Population;
 import com.example.whatseoul.entity.PopulationForecast;
