+++ src/main/java/com/example/whatseoul/service/AccountService.java
@@ -1,7 +1,7 @@
 package com.example.whatseoul.service;
 
 import com.example.whatseoul.entity.User;
-import com.example.whatseoul.respository.user.UserRepository;
+import com.example.whatseoul.repository.user.UserRepository;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
