+++ src/main/java/com/example/whatseoul/service/CultureEventService.java
@@ -3,8 +3,8 @@ package com.example.whatseoul.service;
 import com.example.whatseoul.dto.response.CultureEventDto;
 import com.example.whatseoul.entity.Area;
 import com.example.whatseoul.entity.CultureEvent;
-import com.example.whatseoul.respository.cityData.AreaRepository;
-import com.example.whatseoul.respository.cityData.CulturalEventRepository;
+import com.example.whatseoul.repository.cityData.AreaRepository;
+import com.example.whatseoul.repository.cityData.CulturalEventRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
